requirements.txt
opencv-python
pillow
torch
numpy
transformers
ipython
ultralytics
streamlit
paddleocr
paddlepaddle>=2.5.0
